+++ application/src/test/java/com/ericsson/bss/cassandra/ecchronos/application/TestFileBasedRepairConfiguration.java
@@ -15,7 +15,7 @@
 package com.ericsson.bss.cassandra.ecchronos.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Matchers.eq;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -27,7 +27,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
+import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.context.ApplicationContext;
 
 import com.ericsson.bss.cassandra.ecchronos.application.config.Config;