+++ core/src/test/java/com/ericsson/bss/cassandra/ecchronos/core/repair/state/TestNormalizedBaseRange.java
@@ -21,7 +21,7 @@ import nl.jqno.equalsverifier.EqualsVerifier;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
+import org.mockito.junit.MockitoJUnitRunner;
 
 import java.math.BigInteger;